Add tests for basic testing
3 simple tests that connect to the Facebook server and print data. It
can take access token passed as a command line argument or it can also
use KAccounts (built if KAccounts, AccountsQt and SignOnQt are found) to
obtain the token.

Currently it contains tests for retrieving notifications, posts with
like & comment counts and user either by specified id or the current
logged in user (the user that provided the credentials for the access
token).

At some point I'd like to make these into autotests as well, but it
would require faking the server responses to those KJobs, I'm not quite
sure how to do that yet.